firestore-pgvector/functions/requirements.txt
firebase-functions==0.1.1
firebase-admin==6.2.0
google-cloud-aiplatform==1.28.0
pgvector==0.1.8
cloud-sql-python-connector==1.3.0
asyncpg==0.28.0
google-cloud-tasks==2.13.2